+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI2/_Shared/Styles/XyuiControlStyles.cs
@@ -13,6 +13,7 @@ public static partial class XyuiControlStyles
         var styles = new Styles();
         AddButtonFamily(styles);
         AddInputFamily(styles);
+        Slider(styles);
         AddChoiceControls(styles);
         return styles;
     }